Add some code around removal of bulk_identification_code.txt to prevent error messages

+++ test/Interactive/interactive-test.sh
@@ -11,7 +11,10 @@ checkForRoot
 
 # it is necessasry to remove the bulk_identification_code.txt file for the following tests to work
 # bulk_identification_code.txt is only used by bulk and auto registration
-sudo rm /etc/connectd/bulk_identification_code.txt
+if [ -e /etc/connectd/bulk_identification_code.txt ]; then
+    echo "Deleting /etc/connectd/bulk_identification_code.txt"
+    sudo rm /etc/connectd/bulk_identification_code.txt
+fi
 
 # If run with no following parameter, just run the script interactively
 # while capturing the debug output (e.g. for manual testing).
